refactor: Encounter shortcut
fixes #14285
