+++ src/components/Encounter/EncounterCommandDialog.tsx
@@ -9,10 +9,7 @@ import {
   CommandSeparator,
   CommandShortcut,
 } from "@/components/ui/command";
-import {
-  useEncounterShortcutDisplays,
-  useEncounterShortcuts,
-} from "@/hooks/useEncounterShortcuts";
+import { useShortcutDisplay } from "@/context/ShortcutContext";
 import {
   ArrowBigRight,
   Building2,
@@ -37,6 +34,7 @@ import { useEncounter } from "@/pages/Encounters/utils/EncounterProvider";
 import { EncounterRead } from "@/types/emr/encounter/encounter";
 import questionnaireApi from "@/types/questionnaire/questionnaireApi";
 import { useQuery } from "@tanstack/react-query";
+import { useNavigate } from "raviger";
 import { useTranslation } from "react-i18next";
 
 interface ActionItem {
@@ -65,10 +63,16 @@ export function EncounterCommandDialog({
   onOpenChange,
   trigger,
 }: EncounterCommandDialogProps) {
-  const { canWriteSelectedEncounter, canRestartSelectedEncounter } =
-    useEncounter();
+  const {
+    canWriteSelectedEncounter,
+    canRestartSelectedEncounter,
+    actions,
+    selectedEncounterId,
+    primaryEncounterId,
+  } = useEncounter();
   const { t } = useTranslation();
   const [search, setSearch] = useState("");
+  const navigate = useNavigate();
 
   const questionnaireOptions = useQuestionnaireOptions("encounter_actions");
 
@@ -90,8 +94,7 @@ export function EncounterCommandDialog({
     }
   }, [open]);
 
-  const getShortcutDisplay = useEncounterShortcutDisplays();
-  const { handleAction } = useEncounterShortcuts();
+  const getShortcutDisplay = useShortcutDisplay();
 
   const STORAGE_KEY = "encounter-command-dialog-recent-actions";
   const MAX_RECENT_ACTIONS = 4;
@@ -108,24 +111,86 @@ export function EncounterCommandDialog({
   const [recentActionsState, setRecentActionsState] =
     useState<string[]>(getRecentActions);
 
-  // Handle keyboard shortcut to open command dialog
-  useEffect(() => {
-    const handleOpenCommandDialog = () => {
-      onOpenChange(true);
-    };
+  // Build encounter URL helper
+  const buildEncounterUrl = useCallback(
+    (path: string) => {
+      const currentEncounterIdToUse = primaryEncounterId || encounter.id;
+      const baseUrl = `/facility/${encounter.facility.id}/patient/${encounter.patient.id}/encounter/${currentEncounterIdToUse}${path}`;
 
-    document.addEventListener(
-      "open-encounter-command-dialog",
-      handleOpenCommandDialog,
-    );
+      // Add selectedEncounter parameter if we're viewing a different encounter
+      if (
+        selectedEncounterId &&
+        primaryEncounterId &&
+        selectedEncounterId !== primaryEncounterId
+      ) {
+        const separator = path.includes("?") ? "&" : "?";
+        return `${baseUrl}${separator}selectedEncounter=${selectedEncounterId}`;
+      }
 
-    return () => {
-      document.removeEventListener(
-        "open-encounter-command-dialog",
-        handleOpenCommandDialog,
-      );
-    };
-  }, [onOpenChange]);
+      return baseUrl;
+    },
+    [encounter, selectedEncounterId, primaryEncounterId],
+  );
+
+  // Handle action execution
+  const handleAction = useCallback(
+    (actionId: string) => {
+      const actionHandlers: Record<string, () => void> = {
+        "add-service-request": () =>
+          navigate(buildEncounterUrl("/questionnaire/service_request")),
+        "add-medication-request": () =>
+          navigate(buildEncounterUrl("/questionnaire/medication_request")),
+        "add-allergy": () =>
+          navigate(buildEncounterUrl("/questionnaire/allergy_intolerance")),
+        "add-symptoms": () =>
+          navigate(buildEncounterUrl("/questionnaire/symptom")),
+        "add-diagnosis": () =>
+          navigate(buildEncounterUrl("/questionnaire/diagnosis")),
+        "update-encounter": () =>
+          navigate(buildEncounterUrl("/questionnaire/encounter")),
+        "service-requests": () =>
+          navigate(buildEncounterUrl("/service_requests")),
+        "diagnostic-reports": () =>
+          navigate(buildEncounterUrl("/diagnostic_reports")),
+        "clinical-history": () =>
+          navigate(
+            `/facility/${encounter.facility.id}/patient/${encounter.patient.id}/history/responses?sourceUrl=${encodeURIComponent(
+              buildEncounterUrl("/updates"),
+            )}`,
+          ),
+        "treatment-summary": () =>
+          navigate(buildEncounterUrl("/treatment_summary")),
+        "encounter-overview": () => navigate(buildEncounterUrl("/updates")),
+        plots: () => navigate(buildEncounterUrl("/plots")),
+        observations: () => navigate(buildEncounterUrl("/observations")),
+        medicines: () => navigate(buildEncounterUrl("/medicines")),
+        files: () => navigate(buildEncounterUrl("/files")),
+        notes: () => navigate(buildEncounterUrl("/notes")),
+        devices: () => navigate(buildEncounterUrl("/devices")),
+        consents: () => navigate(buildEncounterUrl("/consents")),
+        "mark-as-completed": () => actions.markAsCompleted(),
+        "assign-location": () => actions.assignLocation(),
+        "view-location-history": () => actions.viewLocationHistory(),
+        "manage-care-team": () => actions.manageCareTeam(),
+        "manage-departments": () => actions.manageDepartments(),
+        dispense: () => actions.dispense(),
+        "restart-encounter": () => actions.restartEncounter(),
+      };
+
+      const handler = actionHandlers[actionId];
+      if (handler) {
+        handler();
+        return;
+      }
+
+      // Handle dynamic questionnaire actions
+      if (actionId.startsWith("questionnaire-")) {
+        const slug = actionId.replace("questionnaire-", "");
+        navigate(buildEncounterUrl(`/questionnaire/${slug}`));
+      }
+    },
+    [navigate, buildEncounterUrl, actions, encounter],
+  );
 
   const addRecentAction = useCallback(
     (actionId: string): void => {
@@ -168,12 +233,6 @@ export function EncounterCommandDialog({
             id: "add-diagnosis",
             label: t("add_diagnosis"),
             shortcut: getShortcutDisplay("add-diagnosis"),
-            icon: <Plus />,
-          },
-          {
-            id: "add-questionnaire",
-            label: t("add_form"),
-            shortcut: getShortcutDisplay("add-questionnaire"),
             icon: <Plus />,
           },
           {
@@ -368,9 +427,10 @@ export function EncounterCommandDialog({
     questionnaires,
     search,
     getShortcutDisplay,
-    isLoading,
     canWriteSelectedEncounter,
     canRestartSelectedEncounter,
+    encounter.encounter_class,
+    encounter?.status,
   ]);
 
   const findRecentActions = useCallback(
