+++ src/config/keyboardShortcuts.json
@@ -131,35 +131,26 @@
   ],
 
   "encounter": [
-    { "key": "a", "action": "add-allergy", "description": "Add Allergy", "when": "canEdit" },
-    { "key": "s", "action": "add-symptoms", "description": "Add Symptoms", "when": "canEdit" },
-    { "key": "d", "action": "add-diagnosis", "description": "Add Diagnosis", "when": "canEdit" },
-    { "key": "f", "action": "add-questionnaire", "description": "New Questionnaire/Form", "when": "canEdit && questionnairesEnabled" },
-    { "key": "u e", "action": "update-encounter", "description": "Update Encounter", "when": "canEdit" },
+    { "key": "a", "action": "add-allergy", "description": "Add Allergy", "when": "always" },
+    { "key": "s", "action": "add-symptoms", "description": "Add Symptoms", "when": "always" },
+    { "key": "d", "action": "add-diagnosis", "description": "Add Diagnosis", "when": "always" },
+    { "key": "r", "action": "add-service-request", "description": "Add Service Request", "when": "always" },
+    { "key": "m", "action": "add-medication-request", "description": "Add Medication", "when": "always" },
+    { "key": "f", "action": "add-questionnaire", "description": "New Questionnaire/Form", "when": "always" },
+    { "key": "shift+u", "action": "update-encounter", "description": "Update Encounter", "when": "always" },
     { "key": "h", "action": "clinical-history", "description": "Clinical History", "when": "always" },
-    { "key": "g g", "action": "encounter-overview", "description": "Go to Overview", "when": "always" },
-    { "key": "g p", "action": "plots", "description": "Go to Plots", "when": "always" },
-    { "key": "g o", "action": "observations", "description": "Go to Observations", "when": "always" },
-    { "key": "g m", "action": "medicines", "description": "Go to Medicines", "when": "always" },
-    { "key": "g f", "action": "files", "description": "Go to Files", "when": "always" },
-    { "key": "g n", "action": "notes", "description": "Go to Notes", "when": "always" },
-    { "key": "g d", "action": "devices", "description": "Go to Devices", "when": "always" },
-    { "key": "g c", "action": "consents", "description": "Go to Consents", "when": "always" },
-    { "key": "g s", "action": "service-requests", "description": "Go to Service Requests", "when": "always" },
-    { "key": "g r", "action": "diagnostic-reports", "description": "Go to Diagnostic Reports", "when": "always" },
     { "key": "cmd+e", "action": "open-command-dialog", "description": "Open Command Dialog", "when": "always" },
     { "key": "ctrl+e", "action": "open-command-dialog", "description": "Open Command Dialog", "when": "always" },
-    { "key": "q 1", "action": "questionnaire-1", "description": "Questionnaire 1", "when": "canEdit && questionnairesEnabled" },
-    { "key": "q 2", "action": "questionnaire-2", "description": "Questionnaire 2", "when": "canEdit && questionnairesEnabled" },
-    { "key": "q 3", "action": "questionnaire-3", "description": "Questionnaire 3", "when": "canEdit && questionnairesEnabled" },
-    { "key": "q 4", "action": "questionnaire-4", "description": "Questionnaire 4", "when": "canEdit && questionnairesEnabled" },
-    { "key": "q 5", "action": "questionnaire-5", "description": "Questionnaire 5", "when": "canEdit && questionnairesEnabled" },
-    { "key": "q 6", "action": "questionnaire-6", "description": "Questionnaire 6", "when": "canEdit && questionnairesEnabled" },
-    { "key": "q 7", "action": "questionnaire-7", "description": "Questionnaire 7", "when": "canEdit && questionnairesEnabled" },
-    { "key": "q 8", "action": "questionnaire-8", "description": "Questionnaire 8", "when": "canEdit && questionnairesEnabled" },
-    { "key": "q 9", "action": "questionnaire-9", "description": "Questionnaire 9", "when": "canEdit && questionnairesEnabled" },
-    { "key": "m c", "action": "mark-as-completed", "description": "Mark as Completed", "when": "canEdit" },
-    { "key": "l", "action": "assign-location", "description": "Assign Location", "when": "canEdit" },
-    { "key": "shift+d", "action": "dispense", "description": "Dispense Medication", "when": "canEdit" }
+    { "key": "q 1", "action": "questionnaire-1", "description": "Questionnaire 1", "when": "always" },
+    { "key": "q 2", "action": "questionnaire-2", "description": "Questionnaire 2", "when": "always" },
+    { "key": "q 3", "action": "questionnaire-3", "description": "Questionnaire 3", "when": "always" },
+    { "key": "q 4", "action": "questionnaire-4", "description": "Questionnaire 4", "when": "always" },
+    { "key": "q 5", "action": "questionnaire-5", "description": "Questionnaire 5", "when": "always" },
+    { "key": "q 6", "action": "questionnaire-6", "description": "Questionnaire 6", "when": "always" },
+    { "key": "q 7", "action": "questionnaire-7", "description": "Questionnaire 7", "when": "always" },
+    { "key": "q 8", "action": "questionnaire-8", "description": "Questionnaire 8", "when": "always" },
+    { "key": "q 9", "action": "questionnaire-9", "description": "Questionnaire 9", "when": "always" },
+    { "key": "shift+c", "action": "mark-as-completed", "description": "Mark as Completed", "when": "always" },
+    { "key": "shift+d", "action": "dispense", "description": "Dispense Medication", "when": "always" }
   ]
 }
