+++ src/pages/Encounters/tabs/overview/quick-actions.tsx
@@ -11,17 +11,12 @@ import {
   TestTubeIcon,
 } from "@/CAREUI/icons/CustomIcons";
 
-import {
-  KeyboardShortcutBadge,
-  ShortcutBadge,
-} from "@/Utils/keyboardShortcutComponents";
+import { ShortcutBadge } from "@/Utils/keyboardShortcutComponents";
 
-import { useEncounterShortcutDisplays } from "@/hooks/useEncounterShortcuts";
 import { FormDialog } from "./FormsDialog";
 
 export const QuickActions = (props: React.ComponentProps<"div">) => {
   const { t } = useTranslation();
-  const getShortcutDisplay = useEncounterShortcutDisplays();
 
   return (
     <div
@@ -31,13 +26,13 @@ export const QuickActions = (props: React.ComponentProps<"div">) => {
       <QuickAction
         icon={<AllergyIcon className="text-red-700" />}
         title={t("allergy")}
-        shortcut={getShortcutDisplay("add-allergy")}
+        actionId="add-allergy"
         href={`questionnaire/allergy_intolerance`}
       />
       <QuickAction
         icon={<TestTubeIcon className="text-pink-700 size-8" />}
         title={t("service_request")}
-        shortcut={getShortcutDisplay("add-service-request")}
+        actionId="add-service-request"
         href={`questionnaire/service_request`}
       />
       <QuickAction
@@ -53,7 +48,7 @@ export const QuickActions = (props: React.ComponentProps<"div">) => {
           <QuickAction
             icon={<HealthWorkerIcon className="text-teal-700" />}
             title={t("forms")}
-            shortcut={getShortcutDisplay("add-questionnaire")}
+            actionId="add-questionnaire"
           />
         }
       />
@@ -64,21 +59,19 @@ export const QuickActions = (props: React.ComponentProps<"div">) => {
 export function QuickAction({
   icon,
   title,
-  shortcut,
-  href,
   actionId,
+  href,
   basePath,
   onClick,
   ...props
 }: {
   icon: React.ReactNode;
   title: string;
-  shortcut?: string;
+  actionId?: string;
   href?: string;
   props?: React.ComponentProps<"div">;
   basePath?: string;
   onClick?: () => void;
-  actionId?: string;
 }) {
   const className =
     "flex-1 flex flex-row md:flex-col gap-1.25 p-1 pb-2 rounded-lg shadow bg-white";
@@ -86,24 +79,14 @@ export function QuickAction({
   if (href) {
     return (
       <Link basePath={basePath} href={href} className={className}>
-        <QuickActionContent
-          icon={icon}
-          title={title}
-          shortcut={shortcut}
-          actionId={actionId}
-        />
+        <QuickActionContent icon={icon} title={title} actionId={actionId} />
       </Link>
     );
   }
 
   return (
     <button className={className} {...props} onClick={onClick}>
-      <QuickActionContent
-        icon={icon}
-        title={title}
-        shortcut={shortcut}
-        actionId={actionId}
-      />
+      <QuickActionContent icon={icon} title={title} actionId={actionId} />
     </button>
   );
 }
@@ -111,18 +94,15 @@ export function QuickAction({
 const QuickActionContent = ({
   icon,
   title,
-  shortcut,
   actionId,
 }: {
   icon: React.ReactNode;
   title: string;
-  shortcut?: string;
   actionId?: string;
 }) => {
   return (
     <>
       <div className="relative flex md:py-3 py-0 rounded-t-md rounded-b-lg md:bg-gray-100 bg-white">
-        <KeyboardShortcutBadge shortcut={shortcut} position="top-right" />
         {actionId && <ShortcutBadge actionId={actionId} position="top-right" />}
         <div className="rounded-xl bg-white md:shadow shadow-none mx-auto items-center flex p-2">
           {icon}
